DO NOT ATTEMPT TO RUN (read comments)
DO NOT ATTEMPT TO RUN:
The code itself is fine but the platformIO config has been set up to use my local copies of the QFSAE libraries.  This is because of many quick fixes that had to be done to bring the whole thing together.  Everything except the modifications to the stf4-can library has been pushed to github, but not platformIO at the time of writing.

st-f4can modifications:
 - disable automatic retransmission