Steering 2023 (#56)
* start with base (ethan-ece-night)

* Revert "start with base (ethan-ece-night)"

This reverts commit d8d089c3d9cd0379bd07dcf043841fb490e15609.

* create

start with base (ethan-ece-night)

* integrate published PIO libraries

* looking pretty good so far

still have to do mode + can + neo pixels

* change "COOL" to "TEMP"

* fix formatting

* fix header comments

* display actual values

* basic errors done

error functions still need to be finished

* save all of my useful pages in the programmers manual

* error stuff

* stale can data error works

* C++

it compiles but will it work?  ngl classes being real boring today.. :100:

* everything works in C++

* Prettified Code!

* DO NOT ATTEMPT TO RUN (read comments)

DO NOT ATTEMPT TO RUN:
The code itself is fine but the platformIO config has been set up to use my local copies of the QFSAE libraries.  This is because of many quick fixes that had to be done to bring the whole thing together.  Everything except the modifications to the stf4-can library has been pushed to github, but not platformIO at the time of writing.

st-f4can modifications:
 - disable automatic retransmission

* debug loop time

* update shifting can addr

* assembly

* remove last local path

* cleanup

* refactor

* rename to tftinfo

* Revert "refactor"

This reverts commit 363cc35c9e7798039a9330318a345676df52cbde.

* refactor

* rename tftinfo

---------

Co-authored-by: Jchisholm204 <Jchisholm204@users.noreply.github.com>